api/src/db/models/chat_room.py
from __future__ import annotations

from datetime import datetime, timezone
from typing import Literal

from beanie import Document
from pydantic import Field


class ChatRoom(Document):
    match_id: str
    participants: list[str]
    type: Literal["dm", "group"]
    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
    updated_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))

    class Settings:
        name = "chat_rooms"

