api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
auth0-fastapi-api>=1.0.0b5
pydantic-settings
pymongo[srv]
beanie
pytest
PyYAML